+++ lib/src/codecs_base.dart
@@ -18,7 +18,6 @@ export 'codecs/utf8.dart';
 export 'core/alphabet.dart';
 export 'core/byte.dart';
 export 'core/codec.dart';
-export 'core/decoder.dart';
-export 'core/encoder.dart';
+export 'core/bit.dart';
 export 'crypt.dart';
 export 'utf8.dart';